refactor(settings): keep language dialog on PreferencesScreen
Move LanguageSelectionDialog out of SettingsScreen so Preferences
does not depend on a leftover internal helper. Drop unused Settings
imports left after the UI/privacy switches moved.
